scheduling changes/improvements

+++ routes/options.js
@@ -12,8 +12,6 @@ router.post('/schedule', function(req, res, error){
 
     var on = '58010301ff00ffffff';
     var off = '58010301ff00000000';
-    var sunrise = new Date(req.body.sunrise);
-    var sunset = new Date(req.body.sunset);
     var setpoint;
 
     var flipSwitch = {
@@ -212,52 +210,75 @@ router.post('/schedule', function(req, res, error){
 
 router.post('/sun', function(req, res, error){
 
+    var on = '58010301ff00ffffff';
+    var off = '58010301ff00000000';
     var setpoint;
+    var gattArgs;
 
-    req.body.onAtSunset ? (setpoint = new Date(req.body.sunset)) : (setpoint = new Date (req.body.sunrise));
-    req.body.onAtSunset ? req.body.device_on = true : req.body.device_on = false;
+    if((new Date(req.body.sunset) < new Date() && req.body.onAtSunset ) || (new Date(req.body.sunrise) < new Date() && req.body.offAtSunrise)){
+        res.send('invalid request');
+    }
 
-    console.log('schedulling non-recurring sunrise/sunset control in options/sun', setpoint);
+    else {
+        if(req.body.offAtSunrise) {
+            //flipSwitch.gattArgs.push(off);
+            gattArgs = call.buildGattargs(req.body.mac, off);
+        }
+        if(req.body.onAtSunset) {
+            //flipSwitch.gattArgs.push(on);
+            gattArgs = call.buildGattargs(req.body.mac, on);
 
-    var job = schedule.scheduleJob(setpoint, function(){
+        }
 
-        var child = spawn('gatttool', flipSwitch.gattArgs);
 
-        child.stdout.on('data', function(data){
+        //console.log('this is still testing: ', gattArgs);
 
-            res.send(data);
+        req.body.onAtSunset ? (setpoint = new Date(req.body.sunset)) : (setpoint = new Date (req.body.sunrise));
+        req.body.onAtSunset ? req.body.device_on = true : req.body.device_on = false;
 
-            child.kill();
+        console.log('schedulling non-recurring sunrise/sunset control in options/sun', setpoint);
+
+        var job = schedule.scheduleJob(setpoint, function(){
+
+            var child = spawn('gatttool', gattArgs);
+
+            child.stdout.on('data', function(data){
+
+                res.send(data);
+
+                child.kill();
+            });
+
+            child.on('exit', function(code){
+                console.log('spawned process ended on exit code: ', code);
+            });
+
         });
 
-        child.on('exit', function(code){
-            console.log('spawned process ended on exit code: ', code);
+        job.on('scheduled',function(date){
+            console.log('job scheduled', date);
         });
 
-    });
+        job.on('run', function(){
+            console.log('my job ran');
 
-    job.on('scheduled',function(date){
-        console.log('job scheduled', date);
-    });
+            pg.connect(connectionString, function(err, client, done){
 
-    job.on('run', function(){
-        console.log('my job ran');
+                var query = client.query("UPDATE devices SET device_on='"+ req.body.device_on +"' where mac='" + req.body.mac + "'", function(error, result){
+                    if(error){console.log('there was an error ', error.detail);}
+                })
 
-        pg.connect(connectionString, function(err, client, done){
+                query.on('end',function(result){
+                    client.end();
+                    //res.send(result);
+                })
 
-            var query = client.query("UPDATE devices SET device_on='"+ req.body.device_on +"' where mac='" + req.body.mac + "'", function(error, result){
-                if(error){console.log('there was an error ', error.detail);}
-            })
-
-            query.on('end',function(result){
-                client.end();
-                //res.send(result);
-            })
-
+            });
         });
-    });
 
-    res.sendStatus(200);
+        res.sendStatus(200);
+
+    }
 
 });
